+++ packages/reactabular/CHANGELOG.md
@@ -1,9 +1,10 @@
-6.1.3 / 2016-10-31
+6.2.0 / 2016-10-31
 ==================
 
 ## reactabular-tree
 
   * Bug fix - Allow `tree.toggleChildren` to work without column `props` defined.
+  * Feature - Add `tree.getImmediateChildren`.
 
 ## reactabular-utils